Imported Upstream version 3.0.2-dfsg

+++ src/VBox/Main/VFSExplorerImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: VFSExplorerImpl.cpp 20963 2009-06-26 09:35:27Z vboxsync $ */
+/* $Id: VFSExplorerImpl.cpp $ */
 /** @file
  *
  * IVFSExplorer COM class implementations.
@@ -362,7 +362,7 @@ HRESULT VFSExplorer::deleteFS(TaskVFSExplorer *aTask)
             if (RT_FAILURE(vrc))
                 throw setError(VBOX_E_FILE_ERROR, tr ("Can't delete file '%s' (%Rrc)"), szPath, vrc);
             if(aTask->progress)
-                aTask->progress->setCurrentOperationProgress(fPercentStep * i);
+                aTask->progress->setCurrentOperationProgress((ULONG)(fPercentStep * i));
         }
     }
     catch(HRESULT aRC)
@@ -488,7 +488,7 @@ HRESULT VFSExplorer::deleteS3(TaskVFSExplorer *aTask)
             if (RT_FAILURE(vrc))
                 throw setError(VBOX_E_FILE_ERROR, tr ("Can't delete file '%s' (%Rrc)"), (*it).c_str(), vrc);
             if(aTask->progress)
-                aTask->progress->setCurrentOperationProgress(fPercentStep * i);
+                aTask->progress->setCurrentOperationProgress((ULONG)(fPercentStep * i));
         }
     }
     catch(HRESULT aRC)
